Add base presenter and baseview Restructure project and add ultis
