+++ app/src/main/AndroidManifest.xml
@@ -11,17 +11,17 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         tools:ignore="GoogleAppIndexingWarning">
-        <activity android:name=".view.SignUpActivity"></activity>
-        <activity android:name=".view.LoginActivity" />
-        <activity android:name=".view.OnboardingActivity" />
-        <activity android:name=".view.SplashScreenActivity">
+        <activity android:name=".view.activity.SignUpActivity"></activity>
+        <activity android:name=".view.activity.LoginActivity" />
+        <activity android:name=".view.activity.OnboardingActivity" />
+        <activity android:name=".view.activity.SplashScreenActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".view.MainActivity" />
+        <activity android:name=".view.activity.MainActivity" />
     </application>
 
 </manifest>
