+++ app/src/main/java/com/example/heaapp/base/BasePresenter.java
@@ -1,0 +1,6 @@
+package com.example.heaapp.base;
+
+public interface BasePresenter<V extends BaseView>  {
+    void attachView(V view);
+    void detachView();
+}
