+++ app/src/main/java/com/example/heaapp/base/BaseView.java
@@ -1,0 +1,7 @@
+package com.example.heaapp.base;
+
+import android.content.Context;
+
+public interface BaseView {
+    Context getContext();
+}
