+++ app/src/main/java/com/example/heaapp/ultis/ultis.java
@@ -1,0 +1,28 @@
+package com.example.heaapp.ultis;
+
+import android.content.Context;
+import android.content.Intent;
+import android.os.Parcelable;
+import android.widget.Toast;
+
+public class ultis<Data> {
+    public static Intent setIntent(Context context, Class destination) {
+
+        Intent intent = new Intent(context, destination);
+        context.startActivity(intent);
+
+        return intent ;
+    }
+
+    public Intent setIntentExtra(Context context, Class destination, String key, Data data) {
+        Intent intent = new Intent(context, destination);
+        intent.putExtra(key, (Parcelable) data);
+        context.startActivity(intent);
+
+        return intent ;
+    }
+
+    public static void showMessage(Context context, String message) {
+        Toast.makeText(context, message, Toast.LENGTH_SHORT).show();
+    }
+}
