+++ app/src/main/java/com/example/heaapp/view/fragment/DashBoardFragment.java
@@ -1,10 +1,6 @@
-package com.example.heaapp.fragment;
+package com.example.heaapp.view.fragment;
 
-import android.content.Context;
-import android.net.Uri;
 import android.os.Bundle;
-
-import androidx.fragment.app.Fragment;
 
 import android.view.LayoutInflater;
 import android.view.View;
@@ -14,19 +10,17 @@ import com.example.heaapp.R;
 import com.example.heaapp.base.BaseFragment;
 
 
-public class HealthInforFragment extends BaseFragment {
+public class DashBoardFragment extends BaseFragment {
 
     @Override
     public BaseFragment provideYourFragment() {
-        return new HealthInforFragment();
-
+        return new DashBoardFragment();
     }
 
     @Override
     public View provideYourFragmentView(LayoutInflater inflater, ViewGroup parent, Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment_health_infor,parent,false);
+
+        View view = inflater.inflate(R.layout.fragment_dash_board,parent,false);
         return view;
     }
-
-
 }
