+++ app/src/main/java/com/example/heaapp/view/fragment/HealthInforFragment.java
@@ -1,31 +1,28 @@
-package com.example.heaapp.fragment;
+package com.example.heaapp.view.fragment;
 
-import android.content.Context;
-import android.net.Uri;
 import android.os.Bundle;
-
-import androidx.fragment.app.Fragment;
 
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.TextView;
 
 import com.example.heaapp.R;
 import com.example.heaapp.base.BaseFragment;
 
 
-public class DashBoardFragment extends BaseFragment {
+public class HealthInforFragment extends BaseFragment {
 
     @Override
     public BaseFragment provideYourFragment() {
-        return new DashBoardFragment();
+        return new HealthInforFragment();
+
     }
 
     @Override
     public View provideYourFragmentView(LayoutInflater inflater, ViewGroup parent, Bundle savedInstanceState) {
-
-        View view = inflater.inflate(R.layout.fragment_dash_board,parent,false);
+        View view = inflater.inflate(R.layout.fragment_health_infor,parent,false);
         return view;
     }
+
+
 }
