+++ app/src/main/java/com/example/heaapp/view/fragment/WorkoutFragment.java
@@ -1,10 +1,6 @@
-package com.example.heaapp.fragment;
+package com.example.heaapp.view.fragment;
 
-import android.content.Context;
-import android.net.Uri;
 import android.os.Bundle;
-
-import androidx.fragment.app.Fragment;
 
 import android.view.LayoutInflater;
 import android.view.View;
