+++ app/src/main/res/layout/activity_login.xml
@@ -7,7 +7,7 @@
     android:orientation="vertical"
     android:fitsSystemWindows="true"
     android:background="@drawable/background_gradient"
-    tools:context=".view.LoginActivity">
+    tools:context=".view.activity.LoginActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
