+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:id="@+id/activity_main_drawer"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="view.MainActivity">
+    tools:context=".view.activity.MainActivity">
 
 
     <LinearLayout
