+++ app/src/main/res/layout/activity_onboarding.xml
@@ -5,7 +5,7 @@
     android:background="#fff"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.OnboardingActivity">
+    tools:context=".view.activity.OnboardingActivity">
 
     <androidx.viewpager.widget.ViewPager
         android:id="@+id/screen_viewpager"
