+++ app/src/main/res/layout/activity_sign_up.xml
@@ -7,7 +7,7 @@
     android:orientation="vertical"
     android:fitsSystemWindows="true"
     android:background="@drawable/background_gradient"
-    tools:context=".view.SignUpActivity">
+    tools:context=".view.activity.SignUpActivity">
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
