+++ app/src/main/res/layout/activity_splash_screen.xml
@@ -5,7 +5,7 @@
     android:id="@+id/activity_splashcreen"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.SplashScreenActivity"
+    tools:context=".view.activity.SplashScreenActivity"
     android:background="@drawable/splashscreen">
 
     <LinearLayout
