+++ app/src/main/res/layout/fragment_health_infor.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".fragment.HealthInforFragment">
+    tools:context=".view.fragment.HealthInforFragment">
 
     <TextView
         android:id="@+id/txtDashboard"
