+++ app/src/main/res/layout/fragment_workout.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".fragment.WorkoutFragment">
+    tools:context=".view.fragment.WorkoutFragment">
 
     <TextView
         android:id="@+id/txtDashboard"
